fix: leave suppression evidence when session-temp exemption mutes churn/W-14
Issue #691: the #688 session-temp exemption silently skipped churn and
W-14 with no event, so stats could not distinguish "guard did not fire"
from "guard was exempted here" — inconsistent with the cooldown path's
pass/skipped evidence pattern.

- Sink the exemption from detect_history_warnings into detect_churn /
  detect_w14: the detectors first compute whether a finding would fire,
  and only then, on temp paths, append a pass/skipped event
  ("[CHURN] suppressed session-temp" / "[W-14] overlap suppressed
  session-temp") instead of the warning. Ordinary temp writes with no
  finding produce zero events, so no constant log overhead.
- Split the inline unit tests of hook_orchestrator_post_edit_history.rs
  into hook_orchestrator_post_edit_history_unit_tests.rs — the evidence
  branch pushed the file past the U-16 800-line ceiling.

Constraint: evidence only on would-have-warned, per the issue's decisive
rationale against per-write logging.
Rejected: unconditional append in the exemption branch — adds constant
log volume on a pure noise-reduction change (the #688 review round
already rejected this).
Confidence: high
Scope-risk: narrow
Verification: cargo test (23 suites; new session_temp_suppression_leaves_
evidence_only_for_real_findings), cargo clippy -D warnings clean,
tests/hooks/test_post_edit_w14.sh 35/35, tests/test_hooks.sh all shards

Fixes #691
